Refactor AccountStatment component: update state management, enhance date handling, and correct routing paths
