+++ Frontend/src/Components/Mainapp/Mainapp.jsx
@@ -80,7 +80,7 @@ import UploadAccount from "./Accounts/Credit/UploadAccount";
 import DairySetup from "./Settings/DairySettings/DairySetup";
 import LedgerList from "./Accounts/LedgerList/LedgerList";
 import MilkColleform from "./Apphome/Appnavviews/Milkcollection/Milkcollform";
-import AccoundStatment from "./Accounts/AccoundStatment/AccoundStatment";
+import AccountStatment from "./Accounts/AccountStatment/AccountStatment";
 
 const Mainapp = () => {
   const dispatch = useDispatch();
@@ -250,9 +250,9 @@ const Mainapp = () => {
             <Route path="accounts/cash" element={<CashCredit />} />
             <Route path="accounts/transfer" element={<TransferCredit />} />
             <Route path="accounts/daybook" element={<DayBook />} />
-            <Route path="accounts/accoundstatment" element={<AccoundStatment />} />
+            <Route path="accounts/statments" element={<AccountStatment />} />
             <Route path="accounts/upload" element={<UploadAccount />} />
-            <Route path="accounts/ledger-list" element={<LedgerList/>} />
+            <Route path="accounts/ledger-list" element={<LedgerList />} />
             {/* master routes */}
             <Route path="master/main-ledger/*" element={<MainLedger />} />
             <Route path="master/sub-ledger/*" element={<SubLedger />} />
