+++ Frontend/src/Components/Mainapp/Mainappnavlinks.jsx
@@ -270,21 +270,21 @@ const Mainappnavlinks = ({ setselected, handleSidebar }) => {
         {
           name: `${t("अकाउंट स्टेटमेंट  ")}`,
           icon: <FaFileInvoice className="icon" />,
-          index: 4.3,
-          path: "accounts/accoundstatment",
+          index: 4.4,
+          path: "accounts/statments",
           role: ["admin", "super_admin", "manager"],
         },
         {
           name: `${t("येणे/देणे खतावणी यादी पाहणे")}`,
           icon: <FaFileInvoice className="icon" />,
-          index: 4.4,
+          index: 4.5,
           path: "accounts/ledger-list",
           role: ["admin", "super_admin", "manager"],
         },
         {
           name: `${t("व्यवहार एक्सेल अपलोड")}`,
           icon: <FaFileInvoice className="icon" />,
-          index: 4.5,
+          index: 4.6,
           path: "accounts/upload",
           role: ["admin", "super_admin", "manager"],
         },
